[fix] price の戻り値を int に修正
関数の戻り値の型と return の型が不一致でも mypy でエラーにならない!!

+++ chapter.6/property.py
@@ -50,7 +50,7 @@ class Book:
         self._discounts = value
 
     @property
-    def price(self) -> float:
+    def price(self) -> int:
         """割引価格."""
         print("call price getter.")
         multi = 100 - self._discounts
